+++ src/core/HierarchicalStorage.cpp
@@ -732,7 +732,7 @@ void __fastcall TRegistryStorage::WriteInt64(const UnicodeString Name, __int64 V
 }
 //---------------------------------------------------------------------------
 void __fastcall TRegistryStorage::WriteBinaryData(const UnicodeString Name,
-  const void * Buffer, int Size)
+  const void * Buffer, size_t Size)
 {
   try
   {
@@ -1035,7 +1035,7 @@ void __fastcall TCustomIniFileStorage::WriteStringRaw(const UnicodeString Name, 
 }
 //---------------------------------------------------------------------------
 void __fastcall TCustomIniFileStorage::WriteBinaryData(const UnicodeString Name,
-  const void * Buffer, int Size)
+  const void * Buffer, size_t Size)
 {
   WriteStringRaw(Name, BytesToHex(RawByteString(static_cast<const char*>(Buffer), Size)));
 }
